Refactor extracting facets
